+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased](https://github.com/nationalarchives/tna-frontend/compare/v0.32.0...HEAD)
 
 ### Added
+
+- If enabled, always show the theme selector in the footer with a message to accept settings cookies if they haven't been
+
 ### Changed
 ### Deprecated
 ### Removed